+++ goldsberry/playtype/_PlayType.py
@@ -1,5 +1,5 @@
 import requests as _requests
-from goldsberry._apiFunc import *
+from goldsberry.apiconvertor import *
 
 class transition:
     def __init__(self, team=False):
